Refactoring of appstate: split into functions
