+++ src/app/TODO.txt
@@ -1,6 +1,5 @@
 p0
 - Экран подтверждения: дабл-рон (нужно собрать и прокинуть данные со стороны апи)
-- Не удается выбрать якухай2 + иццу, проверить якухай2 + саншоку, +доко
 - Починить таймер - переработать его чтобы при суспенде телефона он не останавливался.
 - В интерактивке на странице последней игры выводить дельту рейтинга/очков
 - Подтянуть правки с ноута: В интерактивке нельзя одновременно выбрать тойтой+сананко, иппацу+сананко, возможно что-то еще...
