+++ src/app/primitives/appstate/initials.ts
@@ -1,0 +1,84 @@
+import {
+  AppOutcome,
+  AppOutcomeRon,
+  AppOutcomeTsumo,
+  AppOutcomeDraw,
+  AppOutcomeAbort,
+  AppOutcomeChombo,
+  AppOutcomeMultiRon
+} from '../../interfaces/app';
+import { Outcome as OutcomeType } from '../../interfaces/common';
+import { getFixedFu } from '../yaku-values';
+
+export function initBlankOutcome(outcome: OutcomeType): AppOutcome {
+  let out: AppOutcome;
+  switch (outcome) {
+    case 'ron':
+      const outcomeRon: AppOutcomeRon = {
+        selectedOutcome: 'ron',
+        roundIndex: this._currentRound,
+        loser: null,
+        winner: null,
+        han: 0,
+        fu: 30,
+        possibleFu: getFixedFu([]),
+        yaku: [],
+        riichiBets: [],
+        dora: 0
+      };
+      out = outcomeRon;
+      break;
+    case 'multiron':
+      const outcomeMultiRon: AppOutcomeMultiRon = {
+        selectedOutcome: 'multiron',
+        roundIndex: this._currentRound,
+        loser: null,
+        multiRon: 0,
+        riichiBets: [],
+        wins: {}
+      };
+      out = outcomeMultiRon;
+      break;
+    case 'tsumo':
+      const outcomeTsumo: AppOutcomeTsumo = {
+        selectedOutcome: 'tsumo',
+        roundIndex: this._currentRound,
+        winner: null,
+        han: 0,
+        fu: 30,
+        possibleFu: getFixedFu([]),
+        yaku: [],
+        riichiBets: [],
+        dora: 0
+      };
+      out = outcomeTsumo;
+      break;
+    case 'draw':
+      const outcomeDraw: AppOutcomeDraw = {
+        selectedOutcome: 'draw',
+        roundIndex: this._currentRound,
+        riichiBets: [],
+        tempai: []
+      };
+      out = outcomeDraw;
+      break;
+    case 'abort':
+      const outcomeAbort: AppOutcomeAbort = {
+        selectedOutcome: 'abort',
+        roundIndex: this._currentRound,
+        riichiBets: []
+      };
+      out = outcomeAbort;
+      break;
+    case 'chombo':
+      const outcomeChombo: AppOutcomeChombo = {
+        selectedOutcome: 'chombo',
+        roundIndex: this._currentRound,
+        loser: null
+      };
+      out = outcomeChombo;
+      break;
+  }
+
+  return out;
+}
