+++ src/app/primitives/appstate/timer.ts
@@ -1,0 +1,22 @@
+export class ASTimer {
+  _timeRemaining: number;
+
+  // called from constructor
+  initTimer() {
+    let timer = setInterval(() => {
+      this.decrementTimer();
+      if (!this._timeRemaining) {
+        clearInterval(timer);
+      }
+    }, 1000);
+  }
+
+  getTimeRemaining() {
+    return this._timeRemaining;
+  }
+
+  decrementTimer() {
+    this._timeRemaining--;
+  }
+}
+
